+++ apps/readest-app/src/app/api/share/[token]/og.png/render.tsx
@@ -1,7 +1,7 @@
 import { ImageResponse } from 'next/og';
 import { NextResponse } from 'next/server';
 import { getDownloadSignedUrl } from '@/utils/object';
-import { rejectionToHttp, resolveActiveShare } from '@/libs/share-server';
+import { rejectionToHttp, resolveActiveShare } from '@/libs/shareServer';
 import { SHARE_PRESIGN_TTL_SECONDS } from '@/services/constants';
 
 // JSX renderer for the OG image. Lives in a non-route `.tsx` so the route
