+++ apps/readest-app/src/app/api/share/[token]/revoke/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
 import { createSupabaseAdminClient } from '@/utils/supabase';
 import { validateUserAndToken } from '@/utils/access';
-import { hashShareToken, isValidShareToken } from '@/libs/share-server';
+import { hashShareToken, isValidShareToken } from '@/libs/shareServer';
 
 interface RouteParams {
   params: Promise<{ token: string }>;
