+++ apps/readest-app/src/services/constants.ts
@@ -33,7 +33,7 @@ import { DEFAULT_AI_SETTINGS } from './ai/constants';
 export const DATA_SUBDIR = 'Readest';
 export const LOCAL_BOOKS_SUBDIR = `${DATA_SUBDIR}/Books`;
 export const CLOUD_BOOKS_SUBDIR = `${DATA_SUBDIR}/Books`;
-export const CLOUD_REPLICAS_SUBDIR = `${DATA_SUBDIR}/replicas`;
+export const CLOUD_REPLICAS_SUBDIR = `${DATA_SUBDIR}/Replicas`;
 export const LOCAL_FONTS_SUBDIR = `${DATA_SUBDIR}/Fonts`;
 export const LOCAL_IMAGES_SUBDIR = `${DATA_SUBDIR}/Images`;
 export const LOCAL_DICTIONARIES_SUBDIR = `${DATA_SUBDIR}/Dictionaries`;
